% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filters.R
\name{apply_filters}
\alias{apply_filters}
\title{Apply consecutive filter expressions to a data frame and return results}
\usage{
apply_filters(
  data,
  filters,
  summary_fun = function(data) {
     dplyr::summarise(data)
 },
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame.}

\item{filters}{A list of named quoted expressions.}

\item{summary_fun}{A summary function. Recive a data frame and return a data
frame.}

\item{verbose}{A logical value}
}
\description{
Apply consecutive filter expressions to a data frame and return results
}
\examples{

require(rlang) # summarise, n_distinct, n
require(dplyr) # quo

data <- mtcars

filters <- list(
  `am equals to 1`  = quo(am == 1),
  `carb = 1!!`      = quo(carb == 1),
  `hp less than 90` = quo(hp < 90)
)

summary_fun <- function(data) {
  summarise(
    data,
    n = dplyr::n(),
    cyl_mean = mean(cyl),
    n_unique_cyl = n_distinct(cyl)
    )
}

results <- apply_filters(data, filters, summary_fun)

results

}
