% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_n_labels.R
\name{concentration_max}
\alias{concentration_max}
\title{Get the max proportion}
\usage{
concentration_max(x)
}
\arguments{
\item{x}{The value to get the maximun proportion.}
}
\description{
Get the max proportion
}
\examples{

x <- sample(letters[1:3], prob = 1:3, size = 10000, replace = TRUE)

concentration_max(x)

}
