% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{cor_tidy}
\alias{cor_tidy}
\title{Create a tidy data frame from a correlation (cor) output}
\usage{
cor_tidy(datcor, upper = TRUE)
}
\arguments{
\item{datcor}{A matrix from cor function}

\item{upper}{Ligical. Remove repeated correlations}
}
\description{
Create a tidy data frame from a correlation (cor) output
}
\examples{

data(germancredit, package = "scorecard")

vars <- c("creditability", "duration.in.month", "credit.history", "age.in.years", "purpose")

dat <- germancredit[, vars]

bins <- scorecard::woebin(dat, y = "creditability")

datwoe <- scorecard::woebin_ply(dat, bins)

datwoe <- datwoe[, -c("creditability")]

datcor <- cor(datwoe)

datcor

cor_tidy(datcor)

cor_tidy(datcor, FALSE)

}
