% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-univariate.R
\name{count_distr_max}
\alias{count_distr_max}
\title{Get the max proportion}
\usage{
count_distr_max(x)
}
\arguments{
\item{x}{The value to get the maximun proportion.}
}
\description{
Get the max proportion
}
\examples{

x <- sample(letters[1:3], prob = 1:3, size = 10000, replace = TRUE)

count_distr_max(x)
}
