% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{credit_woe}
\alias{credit_woe}
\title{Credit data in woe form}
\format{
A \code{data frame} with 49,694 observations and 14 variables.
}
\usage{
credit_woe
}
\description{
Credit data in woe form
}
\keyword{datasets}
