% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.R
\name{ct_rules}
\alias{ct_rules}
\title{Get rules from a ctree object}
\usage{
ct_rules(x)
}
\arguments{
\item{x}{a ctree object}
}
\description{
Get rules from a ctree object
}
\examples{

irisct <- partykit::ctree(Species ~ .,data = iris)

ct_rules(irisct)

}
