% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_brute_force}
\alias{featsel_brute_force}
\title{An iterative process to eliminate variables by confidence interval or sign term}
\usage{
featsel_brute_force(model, level = 0.95, verbose = TRUE)
}
\arguments{
\item{model}{model}

\item{level}{the confidence level used to calculate the confidence interval}

\item{verbose}{Default set to TRUE.}
}
\description{
An iterative process to eliminate variables by confidence interval or sign term
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = credit_woe)

m_featsel <- featsel_brute_force(m)

}
