% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_glmnet}
\alias{featsel_glmnet}
\title{Feature selection via glmnet}
\usage{
featsel_glmnet(
  model,
  S = "lambda.1se",
  plot = TRUE,
  seed = 123,
  trace.it = 1,
  type.measue = "auc",
  ...
)
}
\description{
Feature selection via glmnet
}
\examples{


}
