% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_loss_function_permutations}
\alias{featsel_loss_function_permutations}
\title{Shortcut featsel_loss_function_permutations}
\usage{
featsel_loss_function_permutations(
  model,
  stat = c("median", "mean", "min", "q25", "q75", "max"),
  B = 10,
  n_sample = NULL,
  loss_function = DALEX::loss_one_minus_auc,
  verbose = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{model}{model}

\item{stat}{= c("median", "mean", "min", "q25", "q75", "max"),}

\item{B}{Iterations, defaults to 10 for each variable}

\item{n_sample}{n_sample}

\item{loss_function}{Default to DALEX::loss_one_minus_auc}

\item{verbose}{verbose}

\item{plot}{Save plots}

\item{...}{Additional argumentos for DALEX::explain or ingredients::feature_importance}
}
\description{
Shortcut featsel_loss_function_permutations
}
\examples{

if(FALSE){

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = credit_woe)

m_featsel <- featsel_loss_function_permutations(m, stat = "min")

attr(m_featsel, "plots")

}

}
