% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_stepforward}
\alias{featsel_stepforward}
\title{Shortcut for step forward}
\usage{
featsel_stepforward(model, ...)
}
\description{
Shortcut for step forward
}
\examples{

library(scorecard)
data("germancredit")

bins  <-  woebin(germancredit, "creditability")
dtwoe <-  woebin_ply(germancredit, bins)
model <- glm(creditability ~ ., family = binomial(), data = dtwoe)

broom::tidy(model)

model_fssfwd <- featsel_stepforward(model)

broom::tidy(model_fssfwd)


}
