% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-helpers.R
\name{formula_from_reponse_and_predictors_names}
\alias{formula_from_reponse_and_predictors_names}
\title{Get formula from reponse and predictors names}
\usage{
formula_from_reponse_and_predictors_names(response, predictors)
}
\arguments{
\item{response}{response}

\item{predictors}{predictors}
}
\description{
Get formula from reponse and predictors names
}
