% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{gain}
\alias{gain}
\title{Gains}
\usage{
gain(actual, predicted, percents = c(0.1, 0.2, 0.3, 0.4, 0.5))
}
\arguments{
\item{actual}{A binary vector}

\item{predicted}{A numeric vector containing scores or probabilities}

\item{percents}{Values to calculate the gain}
}
\description{
Gains
}
\examples{

N <- 1000
actual <- rbinom(N, size = 1, prob = 0.3)
predicted <- runif(N)

gain(actual, predicted)

}
