% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{gg_woes}
\alias{gg_woes}
\title{Plot woes}
\usage{
gg_woes(
  woes,
  variable = "posprob",
  color_bar = "gray75",
  color_line = "#e22d36",
  ...
)
}
\arguments{
\item{woes}{woes}

\item{variable}{variable}

\item{color_bar}{color_bar}

\item{color_line}{color_line}

\item{...}{...}
}
\description{
Plot woes
}
\examples{

data("woes")

pps <- gg_woes(woes)

if(require(patchwork)){

  purrr::reduce(pps, `+`)  &
     ggplot2::theme(
       axis.text.x = ggplot2::element_text(size = 7),
       axis.text.y = ggplot2::element_text(size = 7)
       )

}

}
