% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_n_labels.R
\name{hhi}
\alias{hhi}
\title{Herfindahl-Hirschman Index}
\usage{
hhi(x)
}
\arguments{
\item{x}{A vector to obtain the HHI.}
}
\description{
Herfindahl-Hirschman Index
}
\examples{

x <- sample(LETTERS[1:10], size = 1000, replace = TRUE, prob = log(1:10))

hhi(x)

plot(table(x), main = hhi_label(hhi(x)))

x <- sample(LETTERS[1:5], size = 1000, replace = TRUE, prob = exp(1:5))

hhi(x)

plot(table(x), main = hhi_label(hhi(x)))

}
