% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_n_labels.R
\name{hhi_label}
\alias{hhi_label}
\title{Get labels for HHI. indexes}
\usage{
hhi_label(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Get labels for HHI. indexes
}
\examples{

hhi_label(c(0.2, 2))

}
