% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmeq}
\alias{hmeq}
\title{hmeq data}
\format{
A \code{data frame} with 5,960 observations and 13 variables.
}
\usage{
hmeq
}
\description{
http://www.creditriskanalytics.net/datasets-private2.html
}
\keyword{datasets}
