% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-predictive.R
\name{iv_label}
\alias{iv_label}
\title{Get labels for Information Values}
\usage{
iv_label(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Get labels for Information Values
}
\examples{

N <- 10000

predicted <- runif(N)

actual <- rbinom(N, size = 1, prob = predicted)

iv_label(information_value(actual, predicted))

predicted <- runif(N)

iv_label(information_value(actual, predicted))
}
