% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-predictive.R
\name{ks}
\alias{ks}
\title{Calculate Kolmogorov-Smirnov statistic}
\usage{
ks(actual, predicted)
}
\arguments{
\item{actual}{A binary vector}

\item{predicted}{A numeric vector containing scores or probabilities}
}
\value{
The KS statistic
}
\description{
More metrics in the Metrics package https://github.com/mfrasco/Metrics
}
\examples{

N <- 10000

predicted <- runif(N)

actual <- rbinom(N, size = 1, prob = predicted)

ks(actual, predicted)

predicted[sample(c(TRUE, FALSE), size = N, prob = c(1, 99), replace = TRUE)] <- NA

ks(actual, predicted)

}
