% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\title{Metrics}
\usage{
metrics(actual, predicted)
}
\arguments{
\item{actual}{A binary vector}

\item{predicted}{A numeric vector containing scores or probabilities}
}
\value{
A data frame with usual and opinionated metrics
}
\description{
Metrics from Metrics package https://github.com/mfrasco/Metrics
}
\examples{

N <- 1000
actual <- rbinom(N, size = 1, prob = 0.3)
predicted <- runif(N)

metrics(actual, predicted)

scorecard::perf_eva(predicted, actual)$binomial_metric$dat


}
