% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{model_summary_variables}
\alias{model_summary_variables}
\alias{model_corr_variables}
\alias{model_vif_variables}
\alias{model_iv_variables}
\title{Get summary of model}
\usage{
model_summary_variables(
  model,
  coef_sign = 1,
  limit_significance = 0.05,
  limit_iv = 0.02,
  limit_corr = 0.6,
  limit_vif = 5
)

model_corr_variables(model)

model_vif_variables(model)

model_iv_variables(model)
}
\arguments{
\item{model}{model}

\item{coef_sign}{Sign to compare estimaes.}

\item{limit_significance}{Limit for Significance.}

\item{limit_iv}{Limit for Information Value.}

\item{limit_corr}{Limit for correlation max between variables.}

\item{limit_vif}{Limit for VIF.}
}
\description{
Get summary of model
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = head(credit_woe, 10000))
m <- featsel_stepforward(m)

model_partials(m)

model_summary_variables(m)

model_corr_variables(m)

model_vif_variables(m)

model_iv_variables(m)

}
