% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-model.R
\name{model_terms_and_ci}
\alias{model_terms_and_ci}
\title{A broom::tidy for glm modelos with terms and confidence intervals}
\usage{
model_terms_and_ci(model, level = 0.95, show_intercept = FALSE)
}
\arguments{
\item{model}{model.}

\item{level}{the confidence level required.}

\item{show_intercept}{A logical value to indicate to show or hide the `(Intercept)`.}
}
\description{
A broom::tidy for glm modelos with terms and confidence intervals
}
