% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-univariate.R
\name{n_growths}
\alias{n_growths}
\title{Number of up/growths}
\usage{
n_growths(x)
}
\arguments{
\item{x}{A numeric vector to calculate the number of growths}
}
\description{
Number of up/growths
}
\examples{

set.seed(123)

x <- round(cumsum(rnorm(24))*100, 0)
x <- abs(x)
x

plot(x, type = "l")

n_growths(x)

n_growths(c(1, 2, 0))

n_growths(c(0))


}
