% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics-helpers.R
\name{plot.model_partials}
\alias{plot.model_partials}
\title{Plot model_partials output}
\usage{
\method{plot}{model_partials}(dfmetrics, ...)
}
\description{
Plot model_partials output
}
\examples{


}
