% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Function to calculate the vector psi given 2 vector of counts}
\usage{
psi(new, old)
}
\arguments{
\item{new}{A vector with the new distribution}

\item{old}{A vector of original distribution}
}
\value{
A list with psi index (value) a label and the table for with counts, percents, woe.
}
\description{
Function to calculate the vector psi given 2 vector of counts
}
\examples{

o <- sample(letters[1:5], size = 2000, prob = 1:5, replace = TRUE)
n <- sample(letters[1:5], size = 1000, prob = 1:5 + 4, replace = TRUE)

table(o)
table(n)

psi(
  table(o),
  table(n)
)

p <- psi(
  table(o),
  table(n)
)

psi_label(sum(p))

}
