% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi_table}
\alias{psi_table}
\title{Function to calculate the vector table given 2 vector of counts}
\usage{
psi_table(old, new)
}
\arguments{
\item{old}{A vector of original distribution}

\item{new}{A vector with the new distribution}
}
\value{
A table with psi index (value) a label and the table for with counts, percents, woe.
}
\description{
Function to calculate the vector table given 2 vector of counts
}
\examples{

o <- factor(sample(letters[1:5], size = 9000, prob = 1:5, replace = TRUE))
n <- factor(sample(letters[1:5], size = 1000, prob = 1:5 + 4, replace = TRUE))

psi_table(o, n)

}
