% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi_vector}
\alias{psi_vector}
\title{Function to calculate the vector psi given 2 vector of counts}
\usage{
psi_vector(old, new)
}
\arguments{
\item{old}{A vector of original count distribution}

\item{new}{A vector with the new count distribution}
}
\value{
A vector with psi index (value) a label and the table for with counts, percents, woe.
}
\description{
Function to calculate the vector psi given 2 vector of counts
}
\examples{

o <- sample(letters[1:5], size = 9000, prob = 1:5, replace = TRUE)
n <- sample(letters[1:5], size = 1000, prob = 1:5 + 4, replace = TRUE)

table(o)
table(n)

psi_vector(
  table(o),
  table(n)
)

p <- psi_vector(
  table(o),
  table(n)
)

p

sum(p)

psi_label(sum(p))

}
