% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-model.R
\name{reponse_and_predictors_names}
\alias{reponse_and_predictors_names}
\title{Get response and predictors names from model}
\usage{
reponse_and_predictors_names(model)
}
\arguments{
\item{model}{model}
}
\description{
Get response and predictors names from model
}
