% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk3-package.R
\name{risk3r-exports}
\alias{risk3r-exports}
\alias{\%>\%}
\title{risk3r exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the risk3r package to avoid listing the magrittr
as Depends of risk3r.
}
