% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{roc_data}
\alias{roc_data}
\alias{ecdf_data}
\title{Auxiliar functions for data plots}
\usage{
roc_data(actual, predicted)

ecdf_data(actual, predicted)
}
\arguments{
\item{actual}{actual}

\item{predicted}{predicted}
}
\description{
Auxiliar functions for data plots
}
\examples{

N <- 10000
predicted <- runif(N)
actual <- rbinom(N, size = 1, prob = predicted)

roc_data(actual, predicted)
ecdf_data(actual, predicted)

}
