% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-univariate.R
\name{rsi}
\alias{rsi}
\alias{cmo}
\title{Relative Strength Index (RSI) and Chande Momentum Oscillator (CMO)}
\usage{
rsi(x)

cmo(x)
}
\arguments{
\item{x}{A numeric vector to calculate the RSI}
}
\description{
Relative Strength Index (RSI) and Chande Momentum Oscillator (CMO)
}
\examples{

set.seed(123)

x <- round(cumsum(rnorm(24)) * 100, 0)
x <- abs(x)
x

plot(x, type = "l")

rsi(x)

# just test
rsi(rev(x))

rsi(x) + rsi(rev(x))
}
