% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_auc.R
\name{step_auc}
\alias{step_auc}
\title{Function to stem model given aux gains}
\usage{
step_auc(model, trace = TRUE)
}
\arguments{
\item{model}{The model to test}

\item{trace}{Show the trace}
}
\description{
Function to stem model given aux gains
}
