% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics-helpers.R
\name{step_forward}
\alias{step_forward}
\title{Shortcut for step forward}
\usage{
step_forward(model, ...)
}
\description{
Shortcut for step forward
}
\examples{

library(scorecard)
data("germancredit")

dt    <-  var_filter(germancredit, "creditability")
bins  <-  woebin(dt, "creditability")
dtwoe <-  woebin_ply(dt, bins)
model <- glm(creditability ~ ., family = binomial(), data = dtwoe)

step_forward(model)

}
