% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-labels.R
\name{vif_label}
\alias{vif_label}
\title{Get labels for VIF values}
\usage{
vif_label(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Get labels for VIF values
}
\examples{

set.seed(123)

vifs <- sample(1:10)

vif_label(vifs)

}
