% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin_cor_iv}
\alias{woebin_cor_iv}
\title{Create a tidy data frame with a correlations and IV}
\usage{
woebin_cor_iv(dt, bins)
}
\arguments{
\item{dt}{A data frame to apply `scorecard::woebin_ply` and calculate correlations.}

\item{bins}{An output from `scorecard::woebin` to create woe variables}
}
\description{
Create a tidy data frame with a correlations and IV
}
\examples{

data(germancredit, package = "scorecard")

vars <- c("creditability", "duration.in.month", "credit.history",
          "purpose", "status.of.existing.checking.account", "property")

dat <- germancredit[, vars]

bins <- scorecard::woebin(dat, y = "creditability", stop_limit = 0.0000001)

woebin_cor_iv(dat, bins)

datcor <- woebin_cor_iv(dat, bins)

library(dplyr)

cor_limit <- 0.15

datcor \%>\%
  filter(variable_1 != variable_2) \%>\%
  mutate(
    cor_conflict = ifelse(abs(cor) > cor_limit, TRUE, FALSE),
    variable_to_remove = ifelse(
      cor_conflict,
      ifelse(iv_variable_1 > iv_variable_2, variable_2, variable_1),
      NA
   )
)

}
