% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin_plot2}
\alias{woebin_plot2}
\title{Alternative plot for woe_bin}
\usage{
woebin_plot2(bins)
}
\arguments{
\item{bins}{An output from `scorecard::woebin` function.}
}
\description{
Alternative plot for woe_bin
}
\examples{

data(germancredit, package = "scorecard")

bins <- scorecard::woebin(
 germancredit,
 y = "creditability",
 x = c("credit.amount", "housing", "duration.in.month"),
 method = "tree"
)

scorecard::woebin_plot(bins)
woebin_plot2(bins)

}
