% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin_ply_min}
\alias{woebin_ply_min}
\title{Minimalistic version of woebin_ply}
\usage{
woebin_ply_min(variable, bin, value = "woe")
}
\arguments{
\item{variable}{A variable to get values asociated to a bin}

\item{bin}{A component (element) of `scorecard::woebin` output}

\item{value}{The value to return (a column of woebin table), defaults to "woe", can be: badprob, bin, etc.}
}
\description{
Minimalistic version of woebin_ply
}
\examples{

data(germancredit, package = "scorecard")

bins <- scorecard::woebin(germancredit, y = "creditability", x = "credit.amount")

variable <- head(germancredit$credit.amount, 10)

bin <- bins$credit.amount

woebin_ply_min(variable, bin)

woebin_ply_min(variable, bin,  value = "badprob")

}
