% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin_summary}
\alias{woebin_summary}
\title{Get scorecard's woe_bin summary}
\usage{
woebin_summary(woebin_obj)
}
\arguments{
\item{woebin_obj}{An output from `scorecard::woebin` function.}
}
\description{
Get scorecard's woe_bin summary
}
\examples{

data(germancredit, package = "scorecard")

woebin_obj <- scorecard::woebin(
 germancredit,
 y = "creditability",
 x = c("credit.amount", "housing", "duration.in.month"),
 method = "tree"
)

woebin_summary(woebin_obj)

}
