% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{woes}
\alias{woes}
\title{woes from credit data}
\format{
A \code{data frame} with 49,694 observations and 14 variables.
}
\usage{
woes
}
\description{
woes from credit data
}
\keyword{datasets}
