% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr-helpers.R
\name{as.list.skim_df}
\alias{as.list.skim_df}
\title{Plot model_partials output}
\usage{
\method{as.list}{skim_df}(x, ...)
}
\arguments{
\item{x}{A skim_df object}

\item{...}{Optional arguments (non used yet).}
}
\description{
Plot model_partials output
}
