% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{bin_plot}
\alias{bin_plot}
\title{Plot one bin from}
\usage{
bin_plot(bin = bins[[3]])
}
\arguments{
\item{bin}{A item from `scorecard::woebin` function.}
}
\description{
Plot one bin from
}
\examples{


data(germancredit, package = "scorecard")

bins <- scorecard::woebin(
  germancredit,
  y = "creditability",
  x = c("credit.amount", "housing", "duration.in.month"),
  method = "tree"
  )

bin_plot(bins[[2]])

}
