% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_plot.R
\name{bin_plot}
\alias{bin_plot}
\title{Plot one bin from}
\usage{
bin_plot(bin = bins[[3]], labels = TRUE)
}
\arguments{
\item{bin}{A item from `scorecard::woebin` function.}

\item{labels}{Logical, show text of values (percentages, bad rates).}
}
\description{
Plot one bin from
}
\examples{

data(germancredit, package = "scorecard")

bins <- scorecard::woebin(
  germancredit,
  y = "creditability",
  x = c("credit.amount", "housing", "duration.in.month"),
  method = "tree"
  )

bin_plot(bins[[2]])

scorecard::woebin_plot(bins[3])

gg <- bin_plot(bins[[2]])


gg


require(ggplot2)
ggplot2::update_geom_defaults("point", list(colour = "black", fill = "white", size = 2, stroke = 2))
ggplot2::update_geom_defaults("point", list(colour = "black", fill = "white", size = 5, stroke = 5))


}
