% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_glmnet}
\alias{featsel_glmnet}
\title{Feature selection via glmnet}
\usage{
featsel_glmnet(
  model,
  S = "lambda.1se",
  plot = TRUE,
  seed = 123,
  trace.it = 1,
  type.measue = "auc",
  ...
)
}
\arguments{
\item{model}{model}

\item{S}{From https://glmnet.stanford.edu/articles/glmnet.html#quick-start: `lambda.min` is
the value of λ that gives minimum mean cross-validated error, while `lambda.1se` is the
value of λ that gives the most regularized model such that the cross-validated error
is within one standard error of the minimum.}

\item{plot}{plot}

\item{seed}{seed}

\item{trace.it}{trace.it p}

\item{type.measue}{type.measue}

\item{...}{Additional argumentos for glmnet::cv.glmnet}
}
\value{
A `glm` object using the selected variables from `glmnet::cv.glmnet`.#'
}
\description{
This is a wrapper for `glmnet::cv.glmnet` and return A `glm` object using
the selected variables from `glmnet::cv.glmnet`.
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = credit_woe)

featsel_glmnet(m)

}
