% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_loss_function_permutations}
\alias{featsel_loss_function_permutations}
\title{Shortcut featsel_loss_function_permutations}
\usage{
featsel_loss_function_permutations(
  model,
  stat = c("median", "mean", "min", "q25", "q75", "max"),
  B = 10,
  n_sample = NULL,
  loss_function = DALEX::loss_one_minus_auc,
  verbose = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{model}{model}

\item{...}{Additional argumentos for stats::step}
}
\description{
Shortcut featsel_loss_function_permutations
}
