% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_loss_function_permutations}
\alias{featsel_loss_function_permutations}
\title{Shortcut featsel_loss_function_permutations}
\usage{
featsel_loss_function_permutations(
  model,
  stat = function(x) quantile(x, 0.25),
  iterations = 100,
  ...
)
}
\arguments{
\item{model}{model}

\item{stat}{= c("median", "mean", "min", "q25", "q75", "max"),}

\item{iterations}{Default 100.}

\item{...}{Additional arguments for celavi::feature_selection}
}
\description{
Shortcut featsel_loss_function_permutations
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = credit_woe)

m_featsel <- featsel_loss_function_permutations(m, stat = min, iterations = 10)

}
