% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_stepforward}
\alias{featsel_stepforward}
\title{Shortcut for step forward}
\usage{
featsel_stepforward(model, ...)
}
\arguments{
\item{model}{model}

\item{...}{Additional argumentos for stats::step}
}
\description{
Shortcut for step forward
}
\examples{

}
