% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-feature-selection.R
\name{featsel_stepforward}
\alias{featsel_stepforward}
\title{Feature selection vis stepwise forward}
\usage{
featsel_stepforward(model, ...)
}
\arguments{
\item{model}{model}

\item{...}{Additional arguments for stats::step}
}
\description{
Feature selection vis stepwise forward
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = credit_woe)

featsel_stepforward(m)

}
