% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplots.R
\name{gg_model_roc}
\alias{gg_model_roc}
\alias{gg_model_ecdf}
\alias{gg_model_coef}
\alias{gg_model_corr}
\alias{gg_model_vif}
\alias{gg_model_importance}
\alias{gg_model_partials}
\title{Plot models}
\usage{
gg_model_roc(model, newdata = NULL, ...)

gg_model_ecdf(model, newdata = NULL, ...)

gg_model_coef(model, level = 0.95, show_intercept = FALSE, ...)

gg_model_corr(model, upper = FALSE, ...)

gg_model_vif(model, colors = c("#3aaf85", "#1b6ca8", "#cd201f"), ...)

gg_model_importance(model, verbose = TRUE, B = 10, ...)

gg_model_partials(model, newdata = NULL, verbose = TRUE, ...)
}
\arguments{
\item{model}{model}

\item{newdata}{newdata}

\item{...}{...}

\item{level}{= 0.95}

\item{show_intercept}{show_intercept}

\item{upper}{upper}

\item{colors}{colors}

\item{verbose}{verbose}

\item{B}{B}
}
\description{
Plot models
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = head(credit_woe, 10000))
m <- featsel_stepforward(m, scale = 5, trace = 0)

dnew <- tail(credit_woe, 10000)

gg_model_roc(m)
gg_model_roc(m, newdata = dnew, size = 2)

gg_model_ecdf(m)
gg_model_ecdf(m, newdata = dnew, size = 2)

gg_model_coef(m)

gg_model_corr(m)

gg_model_vif(m)

gg_model_importance(m)

gg_model_partials(m)

gg_model_coef(m) + ggplot2::coord_flip() + ggplot2::scale_y_discrete(limits = rev)
gg_model_vif(m)  + ggplot2::coord_flip() + ggplot2::scale_x_discrete(limits = rev)
gg_model_corr(m, upper = TRUE)
gg_model_corr(m, upper = TRUE) +
 ggplot2::coord_flip() +
 ggplot2::scale_y_discrete(limits = rev) +
 ggplot2::scale_x_discrete(limits = rev)

}
