% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_n_labels.R
\name{iv_label}
\alias{iv_label}
\title{Get labels for Information Values}
\usage{
iv_label(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Get labels for Information Values
}
\examples{

iv_label(c(0.2, 2))

}
