% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics-predictive.R
\name{metrics}
\alias{metrics}
\title{Metrics}
\usage{
metrics(actual, predicted)
}
\arguments{
\item{actual}{A binary vector}

\item{predicted}{A numeric vector containing scores or probabilities}
}
\value{
A data frame with usual and opinionated metrics
}
\description{
Metrics from Metrics package https://github.com/mfrasco/Metrics
}
\examples{

N <- 10000

predicted <- runif(N)

actual <- rbinom(N, size = 1, prob = predicted)

metrics(actual, predicted)

predicted[sample(c(TRUE, FALSE), size = N, prob = c(1, 99), replace = TRUE)] <- NA

metrics(actual, predicted)
}
