% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics-helpers.R
\name{model_metrics}
\alias{model_metrics}
\title{Calculate predictive metrics for glm models}
\usage{
model_metrics(model, newdata = NULL)
}
\description{
Calculate predictive metrics for glm models
}
\examples{


}
