% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{model_partials}
\alias{model_partials}
\title{Get predictive indicator for partial models given a model}
\usage{
model_partials(model, newdata = NULL, verbose = TRUE)
}
\arguments{
\item{model}{model}

\item{newdata}{Optional data frame}

\item{verbose}{verbose}
}
\description{
Get predictive indicator for partial models given a model
}
\examples{

data("credit_woe")

m <- glm(bad ~ ., family = binomial, data = head(credit_woe, 10000))
m <- featsel_stepforward(m)

model_partials(m)

model_partials(m, newdata = tail(credit_woe, 10000))

}
