% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics.R
\name{model_partials}
\alias{model_partials}
\title{Get predictive indicator for partial models given a model}
\usage{
model_partials(model, newdata = NULL, verbose = TRUE)
}
\arguments{
\item{model}{model}

\item{newdata}{Optional data frame}
}
\description{
Get predictive indicator for partial models given a model
}
\examples{


}
