% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics-helpers.R
\name{model_partials}
\alias{model_partials}
\title{Get predictive indicator for partial models given a model}
\usage{
model_partials(model, newdata = NULL)
}
\description{
Get predictive indicator for partial models given a model
}
\examples{

library(scorecard)
data("germancredit")

dt    <-  var_filter(germancredit, "creditability")
bins  <-  woebin(dt, "creditability")
dtwoe <-  woebin_ply(dt, bins)
model <- glm(creditability ~ ., family = binomial(), data = dtwoe)

modelf <- step_forward(model)

model_partials(modelf)

}
