% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-diagnostics-helpers.R
\name{plot.model_partials}
\alias{plot.model_partials}
\title{Plot model_partials output}
\usage{
\method{plot}{model_partials}(dfmetrics, ...)
}
\description{
Plot model_partials output
}
\examples{

library(scorecard)
data("germancredit")

dt    <-  var_filter(germancredit, "creditability")
bins  <-  woebin(dt, "creditability")
dtwoe <-  woebin_ply(dt, bins)
model <- glm(creditability ~ ., family = binomial(), data = dtwoe)

modelf <- step_forward(model)

dfmetrics <- model_partials(modelf)

plot(dfmetrics)

library(ggplot2)

plot(dfmetrics)

plot(dfmetrics[,-c(4, 5)]) +
  geom_text(
  aes(variable, value + 0.01, label = scales::percent(value, accuracy = 0.1)),
   size = 3
   )


}
