% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin2}
\alias{woebin2}
\title{woebin fix for windows}
\usage{
woebin2(
  dt,
  y,
  x = NULL,
  var_skip = NULL,
  breaks_list = NULL,
  special_values = NULL,
  stop_limit = 0.1,
  count_distr_limit = 0.05,
  bin_num_limit = 8,
  positive = "bad|1",
  no_cores = NULL,
  print_step = 0L,
  method = "tree",
  save_breaks_list = NULL,
  ignore_const_cols = TRUE,
  ignore_datetime_cols = TRUE,
  check_cate_num = TRUE,
  replace_blank_inf = TRUE
)
}
\description{
This function fixes some problem to replicate the cuts given on windows machibes.
}
\details{
See https://github.com/ShichenXie/scorecard/issues/50 for more details.#'
}
\examples{

data(germancredit, package = "scorecard")

bins <- woebin2(
 dt = germancredit,
 y = "creditability",
 x = c("credit.amount", "housing", "duration.in.month", "purpose"),
 method = "tree"
)

bins

}
