% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard-helpers.R
\name{woebin_summary}
\alias{woebin_summary}
\title{Get scorecard's woe_bin summary}
\usage{
woebin_summary(bins, sort = TRUE)
}
\arguments{
\item{bins}{An output from `scorecard::woebin` or `risk3r::woebin2` functions.}

\item{sort}{Sort the data frame by information value. Default TRUE.}
}
\description{
Get scorecard's woe_bin summary
}
\examples{

if(FALSE) {

data(germancredit, package = "scorecard")

bins <- woebin2(
 germancredit,
 y = "creditability",
 # x = c("credit.amount", "housing", "duration.in.month", "purpose"),
 no_cores = 0,
 method = "tree"
)

bin_summary <- woebin_summary(bins)
bin_summary

if(require(dplyr)){

  dplyr::glimpse(bin_summary)

  dplyr::filter(bin_summary, !monotone, !factor)

}

}

}
