% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.R
\name{woebin_summary}
\alias{woebin_summary}
\title{Get scorecard's woe_bin summary}
\usage{
woebin_summary(bins)
}
\arguments{
\item{bins}{An output from `scorecard::woebin` function.}
}
\description{
Get scorecard's woe_bin summary
}
\examples{

data(germancredit, package = "scorecard")

bins <- scorecard::woebin(
 germancredit,
 y = "creditability",
 x = c("credit.amount", "housing", "duration.in.month", "purpose"),
 method = "tree"
)

woebin_summary(bins)

}
